Milestone : Added authorization on a user with security role and regular client.
