--- app/controllers/special_meal_bookings_controller.rb
@@ -1,2 +1,0 @@
-class SpecialMealBookingsController < ApplicationController
-end
