+++ config/routes.rb
@@ -4,9 +4,10 @@
   # root "articles#index"
 Rails.application.routes.draw do
   resources :car_bookings
-  resources :cars
+  
   namespace :api do
     namespace :v1 do
+      resources :cars
       resources :rooms
       resources :room_services
       resources :room_types
