+++ db/seeds.rb
@@ -22,6 +22,7 @@ puts "Seeding Database tables"
 
         user6 = User.create!(username:"ezra",email_address:"ezranjeru@gmail.com",password:"mahalimzuri",password_confirmation:"mahalimzuri",role:"client")
 
+        user7 = User.create!(username:"tasha",email_address:"tashacounts254@gmail.com",password:"mahalimzuri",password_confirmation:"mahalimzuri",role:"admin")
 
     puts "Seeding room type data and rooms data...."
 
@@ -61,7 +62,8 @@ puts "Seeding Database tables"
 
         puts  ".... Seeding special meals booking data for our first user ...."
 
-            specialmealbookings = SpecialMealBooking.create!(user_id:user2.id,room_id:"#{booking1.room_id}",user_room_booking_id:booking1.id,meal_name:"Grilled Fish in Coconut Sauce", meal_type:"Dinner", meal_description:"This dish is a Coastal E/African delicacy that is a loved by almost every family up and down the East Coast, from Mombasa, to Malindi, Lamu and all the way to Zanzibar! A delicious smoky flavourful grilled fish which is then coated with a rich, thick and creamy tamarind-flavoured coconut sauce.This dish is particularly enjoyable with Coconut Rice but can also be had with Mandazi, or Naan and maybe a side of spinach. Once you try this dish, it will definitely count as one of your favourite exotic seafood meals!! YUM!!", number_of_kids:0, number_of_adults:2,booking_date:"2023-05-12T22:00")
+            specialmealbookings = SpecialMealBooking.create!(user_id:user2.id,
+            room_id:"#{booking1.room_id}",user_room_booking_id:booking1.id,meal_name:"Grilled Fish in Coconut Sauce", meal_type:"Dinner", meal_description:"This dish is a Coastal E/African delicacy that is a loved by almost every family up and down the East Coast, from Mombasa, to Malindi, Lamu and all the way to Zanzibar! A delicious smoky flavourful grilled fish which is then coated with a rich, thick and creamy tamarind-flavoured coconut sauce.This dish is particularly enjoyable with Coconut Rice but can also be had with Mandazi, or Naan and maybe a side of spinach. Once you try this dish, it will definitely count as one of your favourite exotic seafood meals!! YUM!!", number_of_kids:0, number_of_adults:2,booking_date:"2023-05-12T22:00")
 
         booking2 = UserRoomBooking.create!(user_id:user6.id,room_id:rand(41..60),number_of_adults:1,number_of_kids:0,booking_date:"2023-05-12T12:00",check_out_date:"2023-05-15T10:00")
